tried to save last best and use it
Tried saving the last best value and stop the gradient from going nuts. It did stop it from happening but didn't improve anything as the direction was still going roughly that way.
Code in this commit makes the next run of the code use the  best result from the last run as starter weights